[net] improve neigh, mtu and pmtu

+++ kernel/networking/internet_layer/icmpv6.h
@@ -32,7 +32,7 @@ typedef struct __attribute__((packed)) {
     uint16_t checksum;
 } icmpv6_hdr_t;
 
-void icmpv6_input(uint16_t ifindex, const uint8_t src_ip[16], const uint8_t dst_ip[16], uint8_t hop_limit, const uint8_t src_mac[6], netpkt_t* pkt);
+void icmpv6_input(uint8_t ifindex, const uint8_t src_ip[16], const uint8_t dst_ip[16], uint8_t hop_limit, const uint8_t src_mac[6], netpkt_t* pkt);
 bool icmpv6_send_on_l2(uint8_t ifindex, const uint8_t dst_ip[16], const uint8_t src_ip[16], const uint8_t dst_mac[6], const void *icmp, uint32_t icmp_len, uint8_t hop_limit);
 
 #ifdef __cplusplus
