+++ kernel/networking/link_layer/eth.h
@@ -24,9 +24,9 @@ uint16_t eth_parse_type(const netpkt_t* pkt);
 bool eth_src(const netpkt_t* pkt, uint8_t out[MAC_ADDR_LEN]);
 bool eth_dst(const netpkt_t* pkt, uint8_t out[MAC_ADDR_LEN]);
 
-bool eth_send_frame_on(uint16_t ifindex, uint16_t ethertype, const uint8_t dst_mac[MAC_ADDR_LEN], netpkt_t* pkt);
+bool eth_send_frame_on(uint8_t ifindex, uint16_t ethertype, const uint8_t dst_mac[MAC_ADDR_LEN], netpkt_t* pkt);
 
-void eth_input(uint16_t ifindex, netpkt_t* pkt);
+void eth_input(uint8_t ifindex, netpkt_t* pkt);
 
 #ifdef __cplusplus
 }
